Refactor PopupMenu and PopupMenuItem
